chore: update generated file names
