+++ frontend/services/identity-manager/account/state.ts
@@ -1,7 +1,7 @@
 import { getUserNumber } from "frontend/services/internet-identity/userNumber"
 import { atom } from "jotai"
 import { atomWithStorage } from "jotai/utils"
-import { AccountResponse } from "../identity_manager"
+import { AccountResponse } from "../identity_manager.did"
 import { ACCOUNT_LOCAL_STORAGE_KEY } from "./constants"
 
 export interface LocalAccount extends Omit<AccountResponse, "anchor"> {
