+++ frontend/services/identity-manager/persona/hooks.ts
@@ -2,7 +2,7 @@ import React from "react"
 import {
   HTTPAccountRequest,
   _SERVICE as _IDENTITY_MANAGER_SERVICE,
-} from "frontend/services/identity-manager/identity_manager"
+} from "frontend/services/identity-manager/identity_manager.did"
 import { useAtom } from "jotai"
 import { personaAtom } from "./state"
 import { isNFIDPersona, isIIPersona } from "./types"
