+++ frontend/services/pub-sub-channel/use-pub-sub-channel.ts
@@ -7,7 +7,7 @@ import {
   Topic,
   _SERVICE,
 } from "frontend/services/pub-sub-channel/pub_sub_channel.did"
-import { idlFactory as pub_sub_channel_idl } from "frontend/services/pub-sub-channel/pub_sub_channel_idl"
+import { idlFactory as pub_sub_channel_idl } from "./pub_sub_channel_idl"
 import { useAuthentication } from "frontend/hooks/use-authentication"
 
 const pubSubCanisterId: string = CONFIG.PUB_SUB_CHANNEL_CANISTER_ID as string
